// src/types/payment.ts
export const platformFeeSchema = z.object({
    fixed: z.unknown(),
    percentage: z.unknown(),
});

export type PlatformFee = z.infer<typeof platformFeeSchema>;

export const refundsSchema = z.object({
    object: z.string(),
    data: z.array(z.unknown()),
    limit: z.number(),
    offset: z.number(),
    total: z.number(),
    location: z.string(),
    order: z.string(),
    from: z.string(),
    to: z.string(),
});
// ...
export const metadataSchema = z.record(z.unknown());

export type Metadata = z.infer<typeof metadataSchema>;

export const sourceSchema = z.object({
    object: z.string(),
    id: z.string(),
    livemode: z.boolean(),
    location: z.string(),
    amount: z.number(),
    barcode: z.unknown().nullable(),
    bank: z.unknown().optional(),
    created_at: z.string(),
    currency: z.string(),
    email: z.unknown(),
    flow: z.string(),
    installment_term: z.unknown(),
    ip: z.unknown().optional(),
    absorption_type: z.unknown().optional(),
    name: z.unknown(),
    mobile_number: z.unknown(),
    phone_number: z.unknown(),
    platform_type: z.unknown().optional(),
    scannable_code: scannableCodeSchema,
    qr_settings: z.unknown().optional(),
    billing: z.unknown(),
    shipping: z.unknown(),
    items: z.array(z.unknown()),
    references: referencesSchema,
    provider_references: providerReferencesSchema.optional(),
    store_id: z.unknown(),
    store_name: z.unknown(),
    terminal_id: z.unknown(),
    type: z.string(),
    zero_interest_installments: z.unknown(),
    charge_status: z.string(),
    receipt_amount: z.unknown().optional(),
    discounts: z.array(z.unknown()).optional(),
    promotion_code: z.unknown(),
});

export type Source = z.infer<typeof sourceSchema>;

export const omiseChargeEntitySchema = z.object({
    object: z.string(),
    id: z.string(),
    location: z.string(),
    amount: z.number(),
    net: z.number(),
    interest: z.number(),
    interest_vat: z.number(),
    refunded_amount: z.number(),
    platform_fee: platformFeeSchema,
    currency: z.string(),
    ip: z.unknown(),
    refunds: refundsSchema,
    link: z.unknown(),
    description: z.unknown(),
    metadata: metadataSchema,
    card: z.unknown(),
    source: sourceSchema,
    schedule: z.unknown(),
    linked_account: z.unknown(),
    customer: z.unknown(),
    dispute: z.unknown(),
    transaction: z.unknown(),
    failure_code: z.unknown(),
    failure_message: z.unknown(),
    status: z.string(),
    authorize_uri: z.string(),
    return_uri: z.unknown(),
    created_at: z.string(),
    paid_at: z.unknown(),
    expires_at: z.string(),
    expired_at: z.unknown(),
    reversed_at: z.unknown(),
    zero_interest_installments: z.boolean(),
    branch: z.unknown(),
    terminal: z.unknown(),
    device: z.unknown(),
    authorized: z.boolean(),
    capturable: z.boolean(),
    capture: z.boolean(),
    disputable: z.boolean(),
    livemode: z.boolean(),
    refundable: z.boolean(),
    reversed: z.boolean(),
    reversible: z.boolean(),
    voided: z.boolean(),
    paid: z.boolean(),
    expired: z.boolean(),
});

export type omiseChargeEntity = z.infer<typeof omiseChargeEntitySchema>;
// ...
export const CardDTOSchema = z.object({
    card_name:  z.string().min(3),
    card_number: z.string().min(3),
    expiration_month: z.number(),
    expiration_year: z.number(),
    security_code: z.string(),
    city: z.string(),
    postal_code: z.string(),
});

export type CardDTOType = z.infer<typeof CardDTOSchema>;

export const contractBookingSchema = z.object({
    id: z.number().int().optional(),
    firstName: z.string().min(3),
    lastName: z.string().min(3),
    email: z.string().email(),
    country: z.string(),
    phoneNumber: z.string().min(10).max(10),
    userId: z.string().optional(),
    created_at: z.union([z.date(), z.string()]).optional(),
    updated_at: z.union([z.date(), z.string()]).optional(),
});

export type contractBookingEntity = z.infer<typeof contractBookingSchema>;

export const BookingByCardDTO = z.object({
    packageId: z.coerce.number().int(),
    childPrice: z.coerce.number().optional(),
    childQty: z.coerce.number().int().optional(),
    adultPrice: z.coerce.number().optional(),
    adultQty: z.coerce.number().int().optional(),
    groupPrice: z.coerce.number().optional(),
    groupQty: z.coerce.number().int().optional(),
    amount: z.coerce.number(),
    additionalDetail: z.string().optional(),
    locationId: z.coerce.number().int().optional(),
    pickup_lat: z.coerce.number(),
    pickup_lgn: z.coerce.number(),
    trip_at: z.union([z.date(), z.string()]),
    pickupLocation: z.string().optional(),
    contractBooking: contractBookingSchema,
    policyAccept: z.boolean(),
    card: CardDTOSchema
});

export type BookingByCardDTOType = z.infer<typeof BookingByCardDTO>;

export const createMobileBankChargeSchema = z.object({
    bank: z.string().optional(),
    bookingId: z.string(),
});

export type createMobileBankChargeType = z.infer<typeof createMobileBankChargeSchema>;